+++ examples/touchpad_gestures.rs
@@ -29,6 +29,9 @@ fn main() {
                         println!("Zoomed out {}", delta);
                     }
                 }
+                WindowEvent::SmartMagnify { .. } => {
+                    println!("Smart zoom");
+                }
                 WindowEvent::TouchpadRotate { delta, .. } => {
                     if delta > 0.0 {
                         println!("Rotated counterclockwise {}", delta);