+++ src/platform_impl/android/mod.rs
@@ -974,6 +974,10 @@ impl Window {
 
     pub fn set_minimized(&self, _minimized: bool) {}
 
+    pub fn is_minimized(&self) -> Option<bool> {
+        None
+    }
+
     pub fn set_maximized(&self, _maximized: bool) {}
 
     pub fn is_maximized(&self) -> bool {