+++ src/platform_impl/orbital/window.rs
@@ -277,6 +277,11 @@ impl Window {
     pub fn set_minimized(&self, _minimized: bool) {}
 
     #[inline]
+    pub fn is_minimized(&self) -> Option<bool> {
+        None
+    }
+
+    #[inline]
     pub fn set_maximized(&self, _maximized: bool) {}
 
     #[inline]